+++ services/soso-web/src/views/DashboardView/LeftSidebarView/components/MemberSidebar.tsx
@@ -1,25 +1,69 @@
-import { MemberCard } from "./MemberCard";
-import { Member } from "./MemberCard";
-import { Sidebar } from "@/components/ui/sidebar";
+// src/views/DashboardView/LeftSidebarView/components/Member_Sidebar.tsx (パスは環境に合わせてください)
 
+import { MemberCard, Member } from "./MemberCard";
+import { 
+    Sidebar,
+    SidebarContent, // ← 追加
+    SidebarGroup, 
+    SidebarHeader, 
+    SidebarGroupLabel, 
+    SidebarGroupContent,
+    SidebarMenu,
+    SidebarMenuItem, // ← 追加
+    useSidebar,      // ← ★重要：これで状態を取得します
+} from "@/components/ui/sidebar";
+import { Avatar, AvatarFallback } from "@/components/ui/avatar"; // ← アバター表示用（なければ単なるdivでもOK）
 
-
-
-
-type Props = {
-  members: Member[];
+type Props =  {
+    members: Member[];
 };
- 
-export function MemberSidebar({ members }: Props) {
-  return (
-    <aside className="flex h-full w-72 flex-col gap-4 rounded-xl bg-slate-50 p-4">
-      <h2 className="text-sm font-semibold text-slate-700">メンバー一覧</h2>
 
-      <div className="mt-2 flex flex-1 flex-col gap-3 overflow-y-auto">
-        {members.map((m) => (
-          <MemberCard key={m.id} member={m} />
-        ))}
-      </div>
-    </aside>
-  );
+export function MemberSidebar({ members }: Props) {
+    // サイドバーの状態を取得 (expanded: 開いている, collapsed: 閉じている)
+    const { state } = useSidebar();
+    const isCollapsed = state === "collapsed";
+
+    return (
+        <Sidebar 
+            collapsible="icon" 
+            side="left"
+            className="border-none bg-slate-50"
+        >
+            <SidebarHeader className={isCollapsed ? "items-center" : ""}>
+                {/* 閉じているときは「一覧」の文字も隠すか、アイコンにする */}
+                {isCollapsed ? (
+                   <span className="text-xs font-bold text-slate-500">List</span>
+                ) : (
+                   <h2 className="text-sm font-semibold text-slate-700">メンバー一覧</h2>
+                )}
+            </SidebarHeader>
+
+            <SidebarContent> {/* ← Contentで囲むのが作法です */}
+                <SidebarGroup>
+                    {!isCollapsed && <SidebarGroupLabel>Application</SidebarGroupLabel>}
+                    <SidebarGroupContent>
+                        <SidebarMenu>
+                            {members.map((member) => (
+                                <SidebarMenuItem key={member.id}>
+                                    {isCollapsed ? (
+                                        // ▼ 【閉じている時】アバター（丸いアイコン）だけ表示
+                                        <div className="flex justify-center py-2">
+                                            <Avatar className="h-8 w-8 cursor-pointer hover:opacity-80">
+                                                <AvatarFallback className="bg-slate-200 text-xs text-slate-600">
+                                                    {member.name.slice(0, 2)} {/* 名前の頭文字 */}
+                                                </AvatarFallback>
+                                            </Avatar>
+                                        </div>
+                                    ) : (
+                                        // ▼ 【開いている時】いつものカードを表示
+                                        <MemberCard member={member} />
+                                    )}
+                                </SidebarMenuItem>
+                            ))}
+                        </SidebarMenu>
+                    </SidebarGroupContent>
+                </SidebarGroup>
+            </SidebarContent>
+        </Sidebar>
+    )
 }
